saas/requirements/local_requirements.txt
asgiref==3.8.1
BareNecessities==0.2.8
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
Django==5.1.1
django-resized==1.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docstring-to-markdown==0.15
et-xmlfile==1.1.0
Faker==30.1.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
imap-tools==1.4.0
jedi==0.19.1
Mail==2.1.0
mailtrap==2.0.1
numpy==1.26.1
oauth2client==4.1.3
onetimepass==1.0.1
openpyxl==3.1.2
pandas==2.1.1
parso==0.8.4
pillow==10.4.0
pluggy==1.5.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
PyJWT==2.9.0
pymongo==4.6.3
pyotp==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
totp==1.3.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.0.7
